+++ CHANGELOG.md
@@ -15,9 +15,11 @@
 - Introduced a setup wizard for onboarding
 - Re-skinned buttons that are available on notifications to the in-house skin
 - Revamp of global Appearances, including the ability to reset, global font, status texture, mute sound, class icon picker, etc.
+- Added a theme selector to the Appearances section.
 - Added a frame to the keystone frame which provides buttons to start a ready check, start a pull timer, and enable or disable automatically starting the key after a pull timer
 - Simplified configuration down to more easy to understand sections, brought together similar configurations that were previously buried into their module's configuration
 - Updated the styling of the Chores tracker frame to be more inline with the current styling; added the ability to configure font and font sizes.
 - Updated some styling in best in slot frames to new styling; more to come.
 - Updated overview page on configuration.
 - Currencies for which you've collected the maximum amount at the time will be highlighted red in the currencies datatext and its tooltip.
+- Added a global error handler to capture errors for later review